i cant think

+++ something-backend/repositories/comment.repo.js
@@ -124,7 +124,7 @@ const getCommentsByPost = async (postId, { limit = 20, cursorCreatedAt = null, c
       COALESCE(v.specialized, 0)::int as vote_specialized,
       u.profile_pic_url as author_profile_pic,
       u.role as author_role,
-      u.rating as author_rating,
+      u.rating::float as author_rating,
       (
         SELECT COUNT(*)::int FROM peer_endorsements pe
         WHERE pe.endorsee_id = c.user_id AND pe.room_id = po.room_id
